% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_item.R
\name{fix_item}
\alias{fix_item}
\title{Fix a reverse coded item}
\usage{
fix_item(value, max)
}
\arguments{
\item{value}{A vector which you'd like reverse coded}

\item{max}{Highest score available to respondents}
}
\value{
A vector of reversed values
}
\description{
Fix a reverse coded item
}
\examples{
fix_item(number_data$q1, 7)
}
