% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_reverse_coding.R
\name{fix_reverse_coding}
\alias{fix_reverse_coding}
\title{Fix multiple reverse coded columns}
\usage{
fix_reverse_coding(df, rc_cols, max)
}
\arguments{
\item{df}{Dataframe with responses where you have reverse coded items}

\item{rc_cols}{Specific columns you want reverse coded}

\item{max}{Highest score available to the respondents}
}
\value{
Dataframe with reverse coded items corrected and replaced
}
\description{
Fix multiple reverse coded columns
}
\examples{
fix_reverse_coding(number_data, c("q1", "q2"), 7)
}
