% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_shortcut.R
\name{sentiment_shortcut}
\alias{sentiment_shortcut}
\title{Create a sentiment analysis dataframe and quick plot}
\usage{
sentiment_shortcut(df, word_column_name, grouping)
}
\arguments{
\item{df}{Dataframe with columns for grouping and the word data}

\item{word_column_name}{Name of column with word responses}

\item{grouping}{Grouping column you want to sort responses by}
}
\value{
Dataframe with tokenized responses attached to a sentiment
value based on afinn dictionary and a plot using smiley
face evaluations of sentiment
}
\description{
Create a sentiment analysis dataframe and quick plot
}
\examples{
sentiment_shortcut(word_data, text, response)
}
